SonTechBot_Project/requirements.txt
kivy
kivymd
kivy-garden.graph
requests
pyodbc
pystray
pillow
pywin32
winshell
pandas
openpyxl
appdirs
getmac
nuitka
zstandard
ordered-set